refactor(computer): move the release bodies (R3, PR-F1e)
The last of the sub-task bodies: what `Effect::SpawnRelease` names, in the
two scopes `sandbox::cleanup` owns. `ReleaseScope::Runtime` is
`release_runtime_resources` — the VMM killed and reaped, the CoW device torn
down, the TAP and its address quarantined, the chroot removed — and
`ReleaseScope::Full` is `release_everything`, which adds what a removal takes
on top: the retained pause overlay, the internal pause checkpoints, and the
working directory. The third scope, `KeepDisk`, moved with pause in PR-F1d
and now shares this module's kill/reap helpers directly.

A pure move apart from `release_everything`, which is a name for the four
statements `remove_sandbox_impl` already ran in that order — worth having,
because `ReleaseScope` has three arms and this file now answers all three.
`remove_sandbox_impl`, the boot failure path, Stop and the expiry loop still
call in, unchanged.

Not carried here: PR-E's gap 5 (idle `Kill` is a non-forced remove today
while the TTL cap is forced). It belongs to the timer path, which is PR-F2b's
— `apply_idle_policy` is what passes `force = false`, and it dies with
`LifecycleTimers`. Nothing diverges in the meantime: the idle timer is armed
only while `Ready`, `begin_removal` refuses a non-forced remove only on
`Running`/`Starting`, and the machine emits the identical teardown for both.
